+++ app/components/Scoreboard.react.js
@@ -4,10 +4,9 @@ import TopicCountItem from 'components/TopicCountItem.react';
 
 export default class Scoreboard extends React.Component {
   render() {
-    const topicListItems = this.props.topics.map((topic) => {
-      return (<TopicCountItem key={topic.get('id')} title={topic.get('text')} count={topic.get('count')}/>);
-    });
-
+    const topicListItems = this.props.topics.toKeyedSeq().map((topic, key) => {
+      return (<TopicCountItem key={key} title={topic.get('text')} count={topic.get('count')}/>);
+    }).toArray();
     return (
       <div className="scoreboard">
         <h3 className="scoreboard__header">Vote count</h3>