+++ app/elements/Header.react.js
@@ -1,0 +1,21 @@
+import React from 'react';
+import Helmet from 'react-helmet';
+
+import config from 'helmconfig.js';
+
+class Header extends React.Component {
+  render() {
+    return (
+      <Helmet
+        title="React Webpack Node"
+        meta={config.meta}
+        link={config.link}
+      />
+    );
+  }
+}
+
+React.renderToString(<Header />);
+let header = Helmet.rewind();
+
+export default header;